refactor(routes): move health HTTP + MCP endpoints into routes/health
Splits src/routes/health/ into mod.rs (wiring), logic.rs (shared
HealthResponse builder), http.rs (GET /health handler), and mcp.rs
(MCP health tool). The MCP tool now builds on the shared logic
instead of duplicating status/uptime/dependencies/version by hand.
