+++ src/routes/health/mod.rs
@@ -1,0 +1,13 @@
+//! Health check: shared logic (`logic.rs`), HTTP handler (`http.rs`), and MCP tool (`mcp.rs`,
+//! declared from `routes::mcp` so it can reach `MoadimMcp`'s private state).
+
+pub(crate) mod logic;
+pub use logic::{DependencyHealth, HealthResponse};
+
+#[path = "http.rs"]
+mod http;
+#[allow(
+    unused_imports,
+    reason = "utoipa's OpenApi derive resolves this hidden __path_health type via crate::routes::health::__path_health, generated by #[utoipa::path] on the re-exported handler below"
+)]
+pub use http::{__path_health, health};
